refactor: small tweaks around validators
