+++ apps/api/src/app/shared/validators/is-boolean-any.validator.ts
@@ -1,7 +1,8 @@
-import { IsIn } from 'class-validator';
+import { IsIn, ValidationOptions } from 'class-validator';
 
 // eslint-disable-next-line @typescript-eslint/naming-convention
-export const IsBooleanAny = () =>
+export const IsBooleanAny = (options: ValidationOptions = {}) =>
   IsIn([true, false, 'true', 'false'], {
     message: '$property must be a boolean value',
+    ...options,
   });
